replace float with bigdecimal for better calculations

+++ app/src/main/java/com/rusmyhal/rates/feature/currencies/data/entity/CurrenciesResponse.kt
@@ -1,8 +1,9 @@
 package com.rusmyhal.rates.feature.currencies.data.entity
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class CurrenciesResponse(
     @SerializedName("base") val baseCurrency: String,
-    @SerializedName("rates") val currenciesRates: Map<String, Float>
+    @SerializedName("rates") val currenciesRates: Map<String, BigDecimal>
 )
